render memory leak re-write part 1

+++ src/main/java/bigtrace/rois/LineTracing3D.java
@@ -105,7 +105,7 @@ public class LineTracing3D implements Roi3D, WritablePolyline
 */
 
 	@Override
-	public void draw(GL3 gl, Matrix4fc pvm, double[] screen_size) {
+	public void draw(GL3 gl, Matrix4fc pvm, int[] screen_size) {
 		
 		
 		float[] colorComp  = new float[]{1.0f, 1.0f, 1.0f, 1.0f};
