+++ src/main/java/bigtrace/rois/PolyLine3D.java
@@ -65,7 +65,7 @@ public class PolyLine3D implements Roi3D, WritablePolyline
 
 
 	@Override
-	public void draw(GL3 gl, Matrix4fc pvm, double[] screen_size) {
+	public void draw(GL3 gl, Matrix4fc pvm, int[] screen_size) {
 		
 		
 		float[] colorComp  = new float[]{1.0f, 1.0f, 1.0f, 1.0f};
