+++ src/main/java/bigtrace/rois/RoiManager3D.java
@@ -295,7 +295,7 @@ public class RoiManager3D extends JPanel implements ListSelectionListener, Actio
 	 }
 	 
 	 /** Draw all ROIS **/
-	 public void draw(GL3 gl, Matrix4fc pvm, double[] screen_size)
+	 public void draw(GL3 gl, Matrix4fc pvm, int[] screen_size)
 	 {
 	       Roi3D roi;
 	       int i;
